Rename project file
